Applying global rules to action chart stats
For example, in book 5 section 166 you lose the use of a shield. This is
taken into account for actual combats, but not for the fake combat
created to parse stats for simply updating the action chart. By making
the fake combat available as part of the action chart object, and
checking for that specifically when processing combat rules, any global
rules that affect stats will be reflected in the action chart.